+++ src/sass/_po-list.scss
@@ -68,6 +68,10 @@ po-line {
         display: inline-block;
     }
 
+    small .row{
+        font-size: 80%;
+    }
+
     .row {
         padding    : $spacer/2 $spacer;
         //border-bottom: 1px solid var(--kt-grey-lighter);
@@ -82,6 +86,31 @@ po-line {
             height    : $po-icon-width;
             min-width : $po-icon-width;
             min-height: $po-icon-width;
+        }
+
+        .small {
+            .predicate-name{
+                font-size: $font-size-sm;
+            }
+
+            .predicate-arg {
+                font-size: $font-size-sm;
+            }
+        }
+
+        .predicate {
+
+            .predicate-name {
+                min-width     : 8rem;
+                font-weight   : 500;
+                text-transform: uppercase;
+            }
+
+            .predicate-arg {
+                font: $font-code;
+                font-size     : $font-smaler;
+                color: var(--kt-grey-dark);
+            }
         }
 
         .predicate {
