+++ test/po-list.html
@@ -1,0 +1,82 @@
+<link rel="import" href="../bower_components/polymer/polymer.html">
+<!-- <link rel="import" href="../bower_components/paper-toolbar/paper-toolbar.html"> -->
+<link rel="import" href="../bower_components/paper-material/paper-material.html">
+<link rel="import" href="../src/components/po_list/po-list.html">
+<link rel="import" href="../src/main-components.html">
+
+<dom-module id="kt-layout">
+    <template>
+
+        <paper-material elevation="1" class="card panel">
+            <po-list small linktosource showgroups filtered-proof-obligations="[[associatedPOs]]"></po-list>
+        </paper-material>
+
+        <paper-material elevation="1" class="card panel">
+            <po-list linktosource showgroups filtered-proof-obligations="[[associatedPOs]]"></po-list>
+        </paper-material>
+
+
+
+    </template>
+    <script>
+
+        Polymer({
+            is: "kt-layout",
+            properties: {
+
+            },
+
+
+            ready: function () {
+
+
+                let ap = [];
+                for (var f = 0; f < 10; f++) {
+                    ap.push({
+                        level: "secondary",
+                        levelLabel:"II",
+                        type: "OP",
+                        label: "some label",
+                        device: "violation",
+                        state: "violation",
+                        predicateArgument: "predicateArgument",
+                        predicate: "predicate",
+                        expression: "expression " + f,
+                        file: "path/to/some/very/deep/file/name" + (f % 2) + ".c",
+                        functionName: "functionName" + (f % 2),
+                        line: 1 + f * 1000,
+                        attr: {
+                            level: "primary",
+                            locationPath: "path/to/some/very/deep/file/name.c/a_long_function_name"
+                        },
+                        renderInfo: {
+                            icon: (f % 2)==0 ? "kt:arrow-forward": "kt:state-discharged-rv",
+                            clazz: ["violation", "s"].join('-').toLowerCase()
+                        }
+
+                    });
+                };
+
+
+                this.associatedPOs = ap;
+
+                const node = this.node;
+                // this.graphHierarchy = {
+                //     node: function (x) {
+                //         return node;
+                //     },
+
+                //     getSuccessors: function (x) {
+                //         return [];
+                //     },
+
+                //     getPredecessors: function (x) {
+                //         return [];
+                //     },
+                // }
+            }
+
+        });
+    </script>
+
+</dom-module>
